clustermesh: drop unnecessary localhost SAN from the admin certificate
The admin certificate is used by the clustermesh-apiserver for client
authentication against the sidecar etcd instance. Etcd leverages the
common name to determine the user name, while SANs are ignored. Hence,
let's just remove it, as not necessary. Additionally, let's do the same
for all the other clustermesh client-auth certs, as by default the CN
is copied into the SAN section as well. This is consistent with the
behavior of the other Cilium's autogeneration methods.

Related: cilium/cilium#32662
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>