Generalize the certificate generation configuration
Currently, certgen has detailed knowledge of the specific certificates
used by Cilium, hence requiring any changes (e.g., introduction of a
new certificate) to be manually applied to this tool as well. Moreover,
it strongly relies on default values, making it difficult to predict
the final output looking at the Cilium's configuration only.

Let's instead introduce a generic configuration, which can be used
to express the list of certificates to be generated, and their
characteristics. The configuration shall be provided in yaml format
through the dedicated flags (--config or --config-file), and an
example is presented in the following:

certs:
- name: foo
  namespace: kube-system
  commonName: foo.cilium.io
  hosts:
  - foo.cilium.io
  - qux.cilium.io
  - 192.0.2.237
  usage:
  - signing
  - key encipherment
  - server auth
  validity: 72h
- name: bar
  ...

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>