renovate: additionally manage go directive in go.mod
By default, renovate does not not update the version specified by the go
directive in the go.mod file, given its "Compatible with this version or
later" meaning [1]. However, certgen is not intended to be imported by
other projects (it has no public API), and it is fine to automatically
bump the go version there, to be able to use all the latest go features.
Still, let's only do minor version updates, to avoid unnecessarily
constraining the patch version as well.

To this end, let's also change the references in various Github actions
with an explicit go version, so that we can make sure of always using
the last available patch version there.

[1]: https://docs.renovatebot.com/modules/manager/gomod/#updating-of-go-mod-and-toolchain-directives

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>