gha: move unit tests execution to the Smoke Test workflow
In preparation for repurposing the Build workflow to perform lint checks
only, let's move the unit tests step to the Smoke Test workflow, which
looks a better fit. To be honest, we don't currently have unit tests,
but let's still keep the step in case they are added in the future.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>