cmd: exit with non-zero code in case of errors
The blamed commit unintentionally changed the behavior in case of
errors, causing the process to return with a zero exit code.
However, this is problematic, as it hides failures and prevents
retries when e.g., executed as a Kubernetes job. Let's restore
the previous behavior, exiting with a non-zero code in that case.

Fixes: 7d1beaa6 ("logging: migrate from logrus to slog")
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>